+++ library/src/main/java/com/getstream/sdk/chat/model/Channel.java
@@ -39,7 +39,6 @@ import com.getstream.sdk.chat.rest.request.MarkReadRequest;
 import com.getstream.sdk.chat.rest.request.ReactionRequest;
 import com.getstream.sdk.chat.rest.request.SendActionRequest;
 import com.getstream.sdk.chat.rest.request.SendEventRequest;
-import com.getstream.sdk.chat.rest.request.SendMessageRequest;
 import com.getstream.sdk.chat.rest.response.ChannelState;
 import com.getstream.sdk.chat.storage.Sync;
 import com.getstream.sdk.chat.storage.converter.DateConverter;
@@ -591,6 +590,31 @@ public class Channel {
         return null;
     }
 
+    /**
+     * countUnread - Count the number of unread messages mentioning the current user
+     *
+     * @return {int} Unread mentions count
+     */
+    public int countUnreadMentions() {
+		Date lastRead = channelState.getReadDateOfChannelLastMessage(client.getUserId());
+        int count = 0;
+        for (Message m : this.channelState.getMessages()) {
+            if (client.getUser().getId().equals(m.getUserId())) {
+                continue;
+            }
+            if (lastRead == null) {
+                count++;
+                continue;
+            }
+            if (m.getCreatedAt().getTime() > lastRead.getTime()) {
+                if (m.getMentionedUsers().indexOf(client.getUser()) != -1) {
+                    count++;
+                }
+            }
+        }
+        return count;
+    }
+
     // region Message
     public void sendMessage(@NonNull Message message,
                             @NonNull MessageCallback callback) {
@@ -600,15 +624,17 @@ public class Channel {
         message.setStatus(getClient().isConnected() ? SENDING : MessageStatus.FAILED);
 
         List<String> mentionedUserIDs = Utils.getMentionedUserIDs(channelState, message.getText());
-        SendMessageRequest request = new SendMessageRequest(message, false, mentionedUserIDs);
-        client.sendMessage(this, request, callback);
+        if (mentionedUserIDs != null && !mentionedUserIDs.isEmpty())
+            message.setMentionedUsersId(mentionedUserIDs);
+        client.sendMessage(this, message, callback);
     }
 
     public void updateMessage(@NonNull Message message,
                               MessageCallback callback) {
         List<String> mentionedUserIDs = Utils.getMentionedUserIDs(channelState, message.getText());
-        SendMessageRequest request = new SendMessageRequest(message, false, mentionedUserIDs);
-        client.updateMessage(message.getId(), request, callback);
+        if (mentionedUserIDs != null && !mentionedUserIDs.isEmpty())
+            message.setMentionedUsersId(mentionedUserIDs);
+        client.updateMessage(message, callback);
     }
 
     public void deleteMessage(@NonNull Message message,