+++ library/src/main/java/com/getstream/sdk/chat/rest/adapter/MessageGsonAdapter.java
@@ -1,6 +1,7 @@
 package com.getstream.sdk.chat.rest.adapter;
 
 import com.getstream.sdk.chat.model.Attachment;
+import com.getstream.sdk.chat.model.ModelType;
 import com.getstream.sdk.chat.model.Reaction;
 import com.getstream.sdk.chat.rest.Message;
 import com.getstream.sdk.chat.rest.User;
@@ -38,11 +39,26 @@ public class MessageGsonAdapter extends TypeAdapter<Message> {
         if (message.getAttachments() != null)
             data.put("attachments", message.getAttachments());
 
-        if (message.getMentionedUsers() != null)
-            data.put("mentioned_users", message.getMentionedUsers());
+        if (message.getMentionedUsersId() != null)
+            data.put("mentioned_users", message.getMentionedUsersId());
 
         if (message.getParentId() != null)
             data.put("parent_id", message.getParentId());
+
+        if (message.getAttachments() != null
+                && !message.getAttachments().isEmpty()) {
+            boolean isGiphy = false;
+            for (Attachment attachment : message.getAttachments()) {
+                if (!isGiphy && attachment.getType().equals(ModelType.attach_giphy))
+                    isGiphy = true;
+            }
+            if (isGiphy){
+                data.put("command", ModelType.attach_giphy);
+                Map<String, String> commandInfo = new HashMap<>();
+                commandInfo.put("name","Giphy");
+                data.put("command_info", commandInfo);
+            }
+        }
 
         TypeAdapter adapter = GsonConverter.Gson().getAdapter(HashMap.class);
         adapter.write(writer, data);