+++ CHANGELOG.md
@@ -7,8 +7,12 @@
 - Improved thread display
 
 ## Oct 14th, 2019 - 2.2.1
+
 - Renamed `FileSendResponse` to `UploadFileResponse`
 - Renamed `SendFileCallback` to `UploadFileCallback`
+- Removed `SendMessageRequest`
+- Updated `sendMessage` and `updateMessage` from `Client`
+- Added devToken function for setUser of Client
 
 ## Oct 10th, 2019 - 2.2.1